+++ packages/administration/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'administration',
-  version: '0.4.4',
+  version: '0.4.5',
   source: '@nuxt-customer-portal/administration',
   dependsOn: ['core', 'ui', 'organizations']
 }