+++ packages/authentication/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'authentication',
-  version: '0.4.4',
+  version: '0.4.5',
   source: '@nuxt-customer-portal/authentication',
   dependsOn: ['core', 'ui']
 }