+++ packages/products/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'products',
-  version: '0.4.4',
+  version: '0.4.5',
   source: '@nuxt-customer-portal/products',
   dependsOn: ['core', 'ui', 'clients'],
   schema: './server/db/schema/products.ts',