+++ packages/planning/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.4.5
+
+### Patch Changes
+
+- Allow providers to save a primary calendar without selecting any additional conflict calendars.
+  - @nuxt-customer-portal/core@0.4.5
+  - @nuxt-customer-portal/products@0.4.5
+  - @nuxt-customer-portal/ui@0.4.5
+
 ## 0.4.4
 
 ### Patch Changes