+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nuxt-customer-portal/ui
 
+## 0.4.5
+
+### Patch Changes
+
+- @nuxt-customer-portal/core@0.4.5
+
 ## 0.4.4
 
 ### Patch Changes